% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{easylayout}
\alias{easylayout}
\title{Layout an igraph object using easylayout's web application.}
\usage{
easylayout(graph, layout = NULL, precompute = TRUE)
}
\arguments{
\item{graph}{An igraph object representing the network to be laid out.}

\item{layout}{Optional initial layout supplied by the user.}

\item{precompute}{Whether or not to skip precomputing the initial layout.}
}
\value{
A two column matrix with XY coordinates and N rows, where N is
the number of vertices in the graph.
}
\description{
The \code{easylayout} function takes an igraph object and serializes`
it into a web application integrated with the IDE’s interface
through a Shiny server. The web application lays out the network by
simulating attraction and repulsion forces. Simulation parameters
can be adjusted in real-time. An editing mode allows moving and rotating
nodes. Once the user finishes tinkering the layout, it is sent back to the
R session to be plotted through popular libraries like ggplot2 or even the
base package itself.
}
\examples{
\dontrun{
library(igraph)
g <- make_ring(10)
g <- easylayout(g)
plot(g)
}
}
